# restapi/internal/api/v1/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Personal Computer Maintenance Management System REST API.",
        "title": "PersonaCMMS API",
        "contact": {
            "email": "greenrivercodelabs@gmail.com"
        },
        "license": {
            "name": "GNU Affero General Public License v3.0",
            "url": "https://spdx.org/licenses/AGPL-3.0-or-later.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/assets/{assetId}/tasks/{taskId}/usage-triggers": {
            "get": {
                "description": "List usage triggers for a task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "usage-triggers"
                ],
                "summary": "List usage triggers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset Id",
                        "name": "assetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Asset Task Id",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.UsageTriggerResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "Create a usage trigger for a task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "usage-triggers"
                ],
                "summary": "Create a usage trigger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset Id",
                        "name": "assetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Asset Task Id",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Usage Trigger object",
                        "name": "usageTrigger",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UsageTriggerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.UsageTriggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/assets/{assetId}/tasks/{taskId}/usage-triggers/{usageTriggerId}": {
            "get": {
                "description": "Get a usage trigger",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "usage-triggers"
                ],
                "summary": "Get a usage trigger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset Id",
                        "name": "assetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Asset Task Id",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Usage Trigger Id",
                        "name": "usageTriggerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.UsageTriggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "Update a usage trigger",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "usage-triggers"
                ],
                "summary": "Update a usage trigger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset Id",
                        "name": "assetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Asset Task Id",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Usage Trigger Id",
                        "name": "usageTriggerId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Usage Trigger object",
                        "name": "usageTrigger",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UsageTriggerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.UsageTriggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a usage trigger",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "usage-triggers"
                ],
                "summary": "Delete a usage trigger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset Id",
                        "name": "assetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Asset Task Id",
                        "name": "taskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Usage Trigger Id",
                        "name": "usageTriggerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.UsageTriggerRequest": {
            "type": "object",
            "required": [
                "quantity",
                "usage_unit"
            ],
            "properties": {
                "quantity": {
                    "type": "integer"
                },
                "usage_unit": {
                    "type": "string"
                }
            }
        },
        "api.UsageTriggerResponse": {
            "type": "object",
            "required": [
                "quantity",
                "usage_unit"
            ],
            "properties": {
                "quantity": {
                    "type": "integer"
                },
                "usage_unit": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}